requirements.txt
django
djangorestframework
environs
psycopg2
dj-database-url
gunicorn